refactor(container): mv TagsBar to unit scope
