+++ src/containers/thread/ReposThread/index.js
@@ -11,7 +11,7 @@ import { ICON_CMD } from '@/config'
 import { THREAD } from '@/constant'
 import { connectStore, buildLog } from '@/utils'
 
-import TagsBar from '@/containers/TagsBar'
+import TagsBar from '@/containers/unit/TagsBar'
 
 import Sticky from '@/components/Sticky'
 import { PublishButton } from '@/components/Buttons'
